feat: update recipe serializers
